requirements.txt
python-telegram-bot==21.6
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
python-dotenv
pytz
faster-whisper